feat: cleanup Orchestrator
